requirements.txt
langchain-openai>=0.0.5
python-dotenv>=1.0.0
pinecone>=6.0.1
boto3>=1.34.0
streamlit>=1.30.0
langchain-core>=0.1.0
langchain>=0.1.0
